// src/app/transfers/models/api/partnerOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssociatedAccount } from './associatedAccount';
import { FilboAccount } from './filboAccount';

export interface PartnerOutputModel { 
    id?: string;
    /**
     * Partner Initials
     */
    twoLetterName?: string;
    name?: string;
    /**
     * True if the partner is a Filbo Customer
     */
    isFilboCustomer?: boolean;
    /**
     * True if any if the accounts in the Accounts list is a Filbo account
     */
    hasFilboAccounts?: boolean;
    /**
     * Two letter country code. RO for Romania
     */
    countryCode?: string;
    /**
     * Country name localized
     */
    countryName?: string;
    /**
     * PF or PJ
     */
    partnerType?: string;
    /**
     * CUI for Romanian Companies
     */
    uniqueCode?: string;
    /**
     * List of accounts mapped with the partner
     */
    accounts?: Array<AssociatedAccount>;
    /**
     * Lista of publicly visible accounts of the partner that are not (yet) added to the accounts list. The list is empty if the partner is not a Filbo Customer
     */
    filboAccounts?: Array<FilboAccount>;
}